# doc/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TRO & LAM API
  version: 1.5.0
  description: |
    Contract frozen from doc/planning/04-data-model.md and 05-api-contracts.md.
    All application routes use /api/v1. Health probes keep the scaffold response
    shape. Provider webhook wire payload and ACK are delegated to the current
    provider documentation; browser redirects never mutate payment state.
servers:
  - url: /api/v1
...
paths:
...
  /products/{id}/reviews:
    parameters: [{ $ref: '#/components/parameters/Id' }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
    get:
      operationId: listPublishedProductReviews
      tags: [catalog]
      x-actors: [guest, customer, staff, admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, default: { $ref: '#/components/responses/Error' } }
...
  /contacts:
    post:
      operationId: createContact
      tags: [support]
      x-actors: [guest, customer]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactCreate' }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '400': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/Error' } }
...
  /account/reviews:
    get:
      operationId: listOwnReviews
      tags: [support]
      x-actors: [customer, staff, admin]
      parameters: [{ in: query, name: status, schema: { $ref: '#/components/schemas/ReviewStatus' } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' } }
    post:
      operationId: createVerifiedReview
      tags: [support]
      x-actors: [customer, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewCreate' } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '409': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
  /account/reviews/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      operationId: updateOwnReview
      tags: [support]
      x-actors: [customer, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewWrite' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /account/tickets:
    get:
      operationId: listOwnTickets
      tags: [support]
      x-actors: [customer, staff, admin]
      parameters: [{ in: query, name: status, schema: { $ref: '#/components/schemas/TicketStatus' } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' } }
  /tickets:
    post:
      operationId: createTicket
      tags: [support]
      x-actors: [customer, guest-order-proof]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TicketCreate' } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
  /orders/{id}/return-requests:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: listOwnedReturnRequests
      tags: [support]
      x-actors: [order-owner, guest-order-proof]
      responses: { '200': { $ref: '#/components/responses/Success' }, '404': { $ref: '#/components/responses/Error' } }
    post:
      operationId: requestReturn
      tags: [support]
      x-actors: [order-owner, guest-order-proof]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnRequestCreate' } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
  /tickets/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getOwnTicket
      tags: [support]
      x-actors: [customer, guest-order-proof, staff, admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, '404': { $ref: '#/components/responses/Error' } }
  /tickets/{id}/messages:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: listTicketMessages
      tags: [support]
      x-actors: [customer, guest-order-proof, staff, admin]
      parameters: [{ in: query, name: cursor, schema: { type: string } }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '404': { $ref: '#/components/responses/Error' } }
    post:
      operationId: createTicketMessage
      tags: [support]
      x-actors: [customer, guest-order-proof, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TicketMessageCreate' } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '404': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
...
  /assistant/handoffs:
    post:
      operationId: createAssistantHandoff
      tags: [assistant]
      x-actors: [guest, customer, order-owner]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AssistantHandoff' } } } }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
  /attachments/uploads:
    post:
      operationId: createAttachmentUpload
      tags: [attachments]
      x-actors: [customer, order-owner, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AttachmentUploadCreate' } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '403': { $ref: '#/components/responses/Error' }, '415': { $ref: '#/components/responses/Error' } }
  /attachments/{id}/finalize:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: finalizeAttachmentUpload
      tags: [attachments]
      x-actors: [customer, order-owner, staff, admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, '404': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
  /attachments/{id}/download:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: downloadOwnedAttachment
      tags: [attachments]
      x-actors: [resource-owner, staff, admin]
      responses: { '302': { description: Authorized redirect to a short-lived private object URL. }, '404': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
...
  /staff/tickets:
    get:
      operationId: listStaffTickets
      tags: [staff]
      x-actors: [staff, admin]
      parameters: [{ in: query, name: status, schema: { $ref: '#/components/schemas/TicketStatus' } }, { in: query, name: assignedTo, schema: { type: string } }, { in: query, name: kind, schema: { $ref: '#/components/schemas/TicketKind' } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
  /staff/tickets/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      operationId: updateStaffTicket
      tags: [staff]
      x-actors: [staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StaffTicketWrite' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /staff/return-requests:
    get:
      operationId: listStaffReturnRequests
      tags: [staff]
      x-actors: [staff, admin]
      parameters: [{ in: query, name: status, schema: { $ref: '#/components/schemas/ReturnStatus' } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
  /staff/return-requests/{id}/decision:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: decideReturnRequest
      tags: [staff]
      x-actors: [staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnDecision' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /staff/return-requests/{id}/inspection:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: inspectReturn
      tags: [staff]
      x-actors: [staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnInspection' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /staff/return-requests/{id}/close:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: closeReturnRequest
      tags: [staff]
      x-actors: [staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReasonVersion' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /staff/contacts:
    get:
      operationId: listStaffContacts
      tags: [staff]
      x-actors: [staff, admin]
      parameters: [{ in: query, name: status, schema: { $ref: '#/components/schemas/ContactStatus' } }, { in: query, name: kind, schema: { $ref: '#/components/schemas/ContactKind' } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
  /staff/contacts/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      operationId: updateStaffContact
      tags: [staff]
      x-actors: [staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StaffContactWrite' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
...
  /admin/reviews:
    get:
      operationId: listAdminReviews
      tags: [admin]
      x-actors: [admin]
      parameters: [{ in: query, name: status, schema: { $ref: '#/components/schemas/ReviewStatus' } }, { in: query, name: productId, schema: { type: string } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
  /admin/reviews/{id}/moderation:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: moderateAdminReview
      tags: [admin]
      x-actors: [admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewModeration' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
...
components:
  parameters:
    Id: { in: path, name: id, required: true, schema: { type: string, minLength: 1 } }
...
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    IdempotencyKey: { in: header, name: Idempotency-Key, required: true, schema: { type: string, minLength: 22, maxLength: 200 }, description: At least 128 bits of client-generated entropy; server scopes it to actor/session and normalized payload. }
  responses:
    Success:
      description: Successful response envelope.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    Created:
      description: Resource created.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    Accepted:
      description: Request persisted for asynchronous delivery/processing.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
...
    Error:
      description: Standardized application error; health and provider webhooks are exceptions.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    SuccessEnvelope: { $ref: './schemas/dtos.schema.json#/$defs/successEnvelope' }
...
    ErrorEnvelope: { $ref: './schemas/dtos.schema.json#/$defs/errorEnvelope' }
...
    ContactCreate: { $ref: './schemas/dtos.schema.json#/$defs/contactCreate' }
...
    ReviewCreate: { $ref: './schemas/dtos.schema.json#/$defs/reviewCreate' }
    TicketCreate: { $ref: './schemas/dtos.schema.json#/$defs/ticketCreate' }
...
    ReturnStatus: { $ref: './schemas/enums.schema.json#/$defs/returnStatus' }
    TicketKind: { $ref: './schemas/enums.schema.json#/$defs/ticketKind' }
    TicketStatus: { $ref: './schemas/enums.schema.json#/$defs/ticketStatus' }
    ContactKind: { $ref: './schemas/enums.schema.json#/$defs/contactKind' }
    ContactStatus: { $ref: './schemas/enums.schema.json#/$defs/contactStatus' }
...
    ReviewStatus: { $ref: './schemas/enums.schema.json#/$defs/reviewStatus' }
...
    ReviewWrite: { $ref: './schemas/dtos.schema.json#/$defs/reviewWrite' }
    ReturnRequestCreate: { $ref: './schemas/dtos.schema.json#/$defs/returnRequestCreate' }
    TicketMessageCreate: { $ref: './schemas/dtos.schema.json#/$defs/ticketMessageCreate' }
...
    AssistantHandoff: { $ref: './schemas/dtos.schema.json#/$defs/assistantHandoff' }
    AttachmentUploadCreate: { $ref: './schemas/dtos.schema.json#/$defs/attachmentUploadCreate' }
...
    StaffTicketWrite: { $ref: './schemas/dtos.schema.json#/$defs/staffTicketWrite' }
    ReturnDecision: { $ref: './schemas/dtos.schema.json#/$defs/returnDecision' }
    ReturnInspection: { $ref: './schemas/dtos.schema.json#/$defs/returnInspection' }
    ReasonVersion: { $ref: './schemas/dtos.schema.json#/$defs/reasonVersion' }
    StaffContactWrite: { $ref: './schemas/dtos.schema.json#/$defs/staffContactWrite' }
...
    ReviewModeration: { $ref: './schemas/dtos.schema.json#/$defs/reviewModeration' }
...